refactor: split src/review.rs into src/review/ sub-modules
Split the 713-line review.rs into a directory module with:
- types.rs (90 lines): all public types and impls
- build.rs (206 lines): artifact construction and proposal logic
- persistence.rs (113 lines): persist/load/latest/render functions
- tests.rs (313 lines): all unit tests
- mod.rs (17 lines): sub-module declarations and re-exports

All pub items re-exported from mod.rs so external crate::review::X
paths remain unchanged. Every sub-module is under 400 lines.
